django-gunicorn/requirements.txt
asgiref==3.5.1
bcrypt==3.2.2
cffi==1.15.0
Django==4.0.4
django-crispy-forms==1.14.0
gunicorn==20.1.0
mysql-connector-python==8.0.29
pika==1.2.1
protobuf==3.20.1
pycparser==2.21
sqlparse==0.4.2
